+++ src/LaravelOptionsServiceProvider.php
@@ -16,6 +16,10 @@ class LaravelOptionsServiceProvider extends BaseServiceProvider
             $this->publishes([
                 __DIR__ . '/../stubs/create_options_table.php.stub' => database_path('migrations/' . date('Y_m_d_His', time()) . '_create_options_table.php'),
             ], 'option-migrations');
+
+            $this->publishes([
+                __DIR__ . '/../config/options.php' => config_path('options.php'),
+            ], 'option-config');
         }
 
         $this->app['options']->boot();
@@ -26,9 +30,12 @@ class LaravelOptionsServiceProvider extends BaseServiceProvider
      */
     public function register()
     {
+        $this->mergeConfigFrom(__DIR__ . '/../config/options.php', 'options');
+
         $this->app->singleton('options', function ($app) {
-            return new Repository($app);
+            return new Repository($app['config']);
         });
+
         $this->app->alias('options', RepositoryContract::class);
     }
 }